test(spotify): add unit tests for SpotifyParser and SpotifyTrackMatcher
- Add SpotifyParserTest to validate URL and URI parsing across supported entity types (track, playlist, album, episode)
- Enforce strict 22-character Base62 ID validation and reject malformed or invalid IDs
- Validate stripping of tracking parameters (?si=) and support for localized URL paths (/intl-en/)
- Verify security boundaries including host-spoofing rejection and rejection of unsupported entities (e.g., artist)
- Add SpotifyTrackMatcherTest covering heuristic selection tiers (official VEVO/- Topic channels, official audio/remaster keywords, and fallback matches)
- Enforce duration bounds filtering ([70%, 130%]) to filter out short previews and full-album compilations
- Cover guard clauses for empty candidate lists, null/blank track titles, and missing artist metadata

+++ src/main/java/com/jagrosh/jmusicbot/spotify/SpotifyBulkLoader.java
@@ -118,7 +118,7 @@ public class SpotifyBulkLoader
         SpotifyTrack firstTrack = result.tracks().get(0);
         String firstTitle = firstTrack.title() != null ? firstTrack.title() : "";
         String firstArtist = firstTrack.artist() != null ? firstTrack.artist() : "";
-        Integer firstDurationMs = firstTrack.durationMs() > 0 ? (int) firstTrack.durationMs() : null;
+        Long firstDurationMs = firstTrack.durationMs() > 0 ? (long) firstTrack.durationMs() : null;
         String firstQuery = (firstTitle + " " + firstArtist).trim();
 
         bot.getPlayerManager().loadItemOrdered(guild, "ytsearch:" + firstQuery,
@@ -281,7 +281,7 @@ public class SpotifyBulkLoader
                             final SpotifyTrack st = result.tracks().get(i);
                             final String sTitle = st.title() != null ? st.title() : "";
                             final String sArtist = st.artist() != null ? st.artist() : "";
-                            final Integer sDurationMs = st.durationMs() > 0 ? (int) st.durationMs() : null;
+                            final Long sDurationMs = st.durationMs() > 0 ? (long) st.durationMs() : null;
                             final String trackQuery = (sTitle + " " + sArtist).trim();
 
                             executor.submit(() -> {
